Crudely hack NATS support into pingmesh
In theory, if this all works, then we can use the `pingmesh.discover` NATS
subject within an account to talk to a NATS server and discover peers, so a
bunch of pingmesh instances in various cloud clusters, where they have routing
to each other, can discover their mutual IPs and keep pinging each other.

In theory.