Remove the addpeer HTTP handler
We learn all our peers via NATS, in an authenticated namespace, so we don't
need to let anyone who can connect to us add peers.

Leaving the /v1/quit in place and the lists of peers.  There's an information
disclosure issue in having these, but nothing not discoverable by resolving
public DNS from a few places.